move: jpn_map from util to visualization
