+++ covsirphy/__init__.py
@@ -11,7 +11,6 @@ except ImportError:
 from covsirphy.__version__ import __version__
 # util
 from covsirphy.util.plotting import line_plot, line_plot_multiple
-from covsirphy.util.map import jpn_map
 from covsirphy.util.optimize import Optimizer
 from covsirphy.util.stopwatch import StopWatch
 from covsirphy.util.error import deprecate
@@ -23,6 +22,7 @@ from covsirphy.util.term import Term, Word
 # visualization
 from covsirphy.visualization.vbase import VisualizeBase
 from covsirphy.visualization.colored_map import ColoredMap
+from covsirphy.visualization.japan_map import jpn_map
 # cleaning
 from covsirphy.cleaning.cbase import CleaningBase
 from covsirphy.cleaning.jhu_data import JHUData
